pyproject.toml
[project]
name = "lunchmoney-assets"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "camoufox[geoip]>=0.4.11",
    "playwright>=1.50.0",
    "python-decouple-typed>=3.11.0",
]

[dependency-groups]
dev = [
    "pudb>=2024.1.3",
]
